create ChooseSeats component
