+++ src/app/(site)/tickets/page.tsx
@@ -4,6 +4,13 @@ import { FC, useState, useEffect } from 'react';
 import Link from 'next/link';
 import TicketMovieInfo from '@/app/components/TicketMovieInfo';
 import TicketCount from '@/app/components/TicketCount';
+import ChooseSeats from '@/app/components/ChooseSeats';
+
+export enum Seat {
+  available,
+  booked,
+  choosed,
+}
 
 const TicketsPage: FC = () => {
   useEffect(() => {
@@ -36,12 +43,6 @@ const TicketsPage: FC = () => {
     fetchData();
   }, []);
 
-  enum Seat {
-    available,
-    booked,
-    choosed,
-  }
-
   const [ticketCount, setTicketCount] = useState(0);
   const [seatsArray, setSeatsArray] = useState<
     [] | { status: Seat; seatNumber: number }[]
@@ -60,78 +61,11 @@ const TicketsPage: FC = () => {
         onClickPlus={() => setTicketCount(ticketCount + 1)}
       />
 
-      <h2 className="header-two">Välj platser</h2>
-
-      <div className="movie-screen-seat-container">
-        <div className="movie-screen-container">
-          <div className="movie-screen-left"></div>
-          <div className="movie-screen-middle"></div>
-          <div className="movie-screen-right"></div>
-        </div>
-
-        <div className="seat-container">
-          {seatsArray.map((seat) => {
-            return (
-              <div
-                key={seat.seatNumber}
-                onClick={() => {
-                  let num = 0;
-                  for (let obj of seatsArray) {
-                    if (obj.status == Seat.choosed) {
-                      num++;
-                    }
-                  }
-                  const newArray = seatsArray.map((item) => {
-                    if (item.seatNumber == seat.seatNumber) {
-                      if (item.status == Seat.available) {
-                        if (num < ticketCount) {
-                          return {
-                            status: Seat.choosed,
-                            seatNumber: seat.seatNumber,
-                          };
-                        } else {
-                          return {
-                            status: Seat.available,
-                            seatNumber: seat.seatNumber,
-                          };
-                        }
-                      } else if (item.status == Seat.choosed) {
-                        return {
-                          status: Seat.available,
-                          seatNumber: seat.seatNumber,
-                        };
-                      } else {
-                        return item;
-                      }
-                    } else {
-                      return item;
-                    }
-                  });
-                  setSeatsArray(newArray);
-                }}
-                className={
-                  seat.status == Seat.booked
-                    ? 'seat-not-available'
-                    : seat.status == Seat.choosed
-                    ? 'seat-choosed'
-                    : 'seat-available'
-                }
-              >
-                {/* {seat.seatNumber} */}
-              </div>
-            );
-          })}
-        </div>
-      </div>
-
-      <div className="choose-seat-container">
-        <div className="choosed-seat"></div>
-        <p>Vald plats</p>
-      </div>
-      <div className="choose-seat-container">
-        <div className="booked-seat"></div>
-        <p>Upptagen</p>
-      </div>
+      <ChooseSeats
+        seatsArray={seatsArray}
+        ticketCount={ticketCount}
+        onSetSeatsArray={(newArray) => setSeatsArray(newArray)}
+      />
 
       <div className="divider"></div>
